backend/requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
